[FEATURE]canchat Implemention

+++ backend/src/services/block.service.ts
@@ -1,4 +1,8 @@
 import { blockRepository } from '../repositories/block.repository.js';
+import { likeRepository } from '../repositories/like.repository.js';
+import { conversationRepository } from '../repositories/conversation.repository.js';
+import { reportRepository } from '../repositories/report.repository.js';
+import { notificationService } from './notification.service.js';
 
 export const blockService = {
   blockUser: async (blockerId: number, blockedId: number) => {
@@ -15,6 +19,23 @@ export const blockService = {
 
     // Create the block
     await blockRepository.createBlock(blockerId, blockedId);
+
+    // Remove likes in both directions to break any match
+    await Promise.all([
+      likeRepository.removeLike(blockerId, blockedId),
+      likeRepository.removeLike(blockedId, blockerId),
+      conversationRepository.removeConversation(blockerId, blockedId),
+    ])
+
+    // Remove notifications in both directions
+    await Promise.all([
+      notificationService.deleteNotification(blockerId, blockedId, "LIKE"),
+      notificationService.deleteNotification(blockedId, blockerId, "LIKE"),
+      notificationService.deleteNotification(blockerId, blockedId, "MATCH"),
+      notificationService.deleteNotification(blockedId, blockerId, "MATCH"),
+      notificationService.deleteNotification(blockerId, blockedId, "VIEW"),
+      notificationService.deleteNotification(blockedId, blockerId, "VIEW"),
+    ]);
 
     return {
       success: true,
@@ -34,6 +55,12 @@ export const blockService = {
       throw new Error('User is not blocked');
     }
 
+    // const isReported = await reportRepository.isReported(blockerId, blockedId);
+    const blockerReportedBlocked = await reportRepository.checkReportExists(blockerId, blockedId);
+    if (blockerReportedBlocked) {
+      throw new Error('Cannot unblock a reported user');
+    }
+    
     // Remove the block
     await blockRepository.removeBlock(blockerId, blockedId);
